refactor: 볼과 스트라이크 개수를 확인하는 기능을 외부 Count class로 분리하여 App.js에서 불러오도록 수정
